refactor: namingof auth stuff
