+++ src/actions/users/update.ts
@@ -3,7 +3,7 @@ import { updateUserSchema, userRegisterSchema } from './schema'
 import { createZodActionError, createActionError } from '@/actions/error'
 import { updateUser } from '@/server/users/update'
 import { registerUser } from '@/server/auth/credentials'
-import { getUser } from '@/auth/user'
+import { getUser } from '@/auth/getUser'
 import type { ActionReturn } from '@/actions/Types'
 import type { User } from '@prisma/client'
 import type { UpdateUserSchemaType } from './schema'
