+++ src/app/components/NavBar/MobileNavBar.tsx
@@ -3,7 +3,7 @@ import styles from './MobileNavBar.module.scss'
 import Menu from './Menu'
 import SpecialCmsImage from '@/components/Cms/CmsImage/SpecialCmsImage'
 import EditModeSwitch from '@/components/EditModeSwitch/EditModeSwitch'
-import { getUser } from '@/auth/user'
+import { getUser } from '@/auth/getUser'
 import Link from 'next/link'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faBars } from '@fortawesome/free-solid-svg-icons'
