+++ src/app/users/[username]/page.tsx
@@ -1,5 +1,5 @@
 import prisma from '@/prisma'
-import { getUser } from '@/auth/user'
+import { getUser } from '@/auth/getUser'
 import { notFound } from 'next/navigation'
 import Link from 'next/link'
 import { v4 as uuid } from 'uuid'
