+++ src/hooks/useEditing.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import { useUser } from "@/auth/client"
+import { useUser } from "@/auth/useUser"
 
 export default function useEditing() {
